# public/api-docs.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Orchids Project Command Center API",
    "description": "Code Style Policeman (CSP) - Team project command center for monitoring development workflow, communication, and code quality heuristics.",
    "version": "1.0.0",
    "contact": {
      "name": "CSP Team"
    }
  },
  "servers": [
    {
      "url": "http://localhost:3000",
      "description": "Development"
    }
  ],
  "paths": {
...
    "/api/workspaces": {
      "get": {
        "summary": "List user's workspaces",
        "tags": ["Workspaces"],
        "security": [{ "bearerAuth": [] }],
        "responses": {
          "200": {
            "description": "List of workspaces",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "workspaces": {
                      "type": "array",
                      "items": { "$ref": "#/components/schemas/Workspace" }
                    }
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "summary": "Create a new workspace",
        "tags": ["Workspaces"],
        "security": [{ "bearerAuth": [] }],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["name"],
                "properties": {
                  "name": { "type": "string", "minLength": 1, "maxLength": 100 },
                  "description": { "type": "string", "maxLength": 500 },
                  "github_repo_url": { "type": "string", "format": "uri" },
                  "github_repo_owner": { "type": "string" },
                  "github_repo_name": { "type": "string" }
                }
              }
            }
          }
        },
        "responses": {
          "201": { "description": "Workspace created" },
          "400": { "$ref": "#/components/responses/ValidationError" }
        }
      }
    },
...
    "/api/workspaces/{workspaceId}/alerts": {
      "get": {
        "summary": "List alerts",
        "tags": ["Alerts"],
        "security": [{ "bearerAuth": [] }],
        "parameters": [{ "$ref": "#/components/parameters/workspaceId" }],
        "responses": {
          "200": { "description": "List of alerts" }
        }
      },
      "patch": {
        "summary": "Resolve an alert",
        "tags": ["Alerts"],
        "security": [{ "bearerAuth": [] }],
        "parameters": [{ "$ref": "#/components/parameters/workspaceId" }],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["alert_id"],
                "properties": {
                  "alert_id": { "type": "string", "format": "uuid" }
                }
              }
            }
          }
        },
        "responses": {
          "200": { "description": "Alert resolved" }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
    "parameters": {
      "workspaceId": {
        "name": "workspaceId",
        "in": "path",
        "required": true,
        "schema": { "type": "string", "format": "uuid" }
      }
    },
    "schemas": {
...
      "Workspace": {
        "type": "object",
        "properties": {
          "id": { "type": "string", "format": "uuid" },
          "name": { "type": "string" },
          "description": { "type": "string", "nullable": true },
          "github_repo_url": { "type": "string", "nullable": true },
          "created_at": { "type": "string", "format": "date-time" }
        }
      },
...
    },
    "responses": {
      "ValidationError": {
        "description": "Validation error",
        "content": {
          "application/json": {
            "schema": {
              "type": "object",
              "properties": {
                "error": { "type": "string" }
              }
            }
          }
        }
      },
...
    }
  }
}